Add smart floor tile tool to map editor
The Grid class is used to keep track of how the floor / walls are laid out on the map.
You can modify the floor tiles of the grid with icons under the new ADVANCED tab in the map editor.
Every time that the the grid is modified, some local corrections are made (updating the wall types, adding/removing entities to the world)

When a map is loaded from file, we attempt to build the grid from the floor tiles in the game state.
However this is crazy expensive for large maps and will need to be improved.